requirements.txt
numpy
scipy
scikit-learn
opencv-python
Pillow==2.2.1
matplotlib
